Extract components and functions to separate files
